+++ includes/cli/templates/block-dynamic/index.js
@@ -1,14 +1,14 @@
 /**
- * template-item-title block main entry point
+ * template-title block main entry point
  *
  * This file orchestrates all the script files
- * together to create a WordPress child block.
+ * together to create a WordPress block.
  */
 
 /**
  * Import dependencies
  */
-import Icon from '../../../icons/general.inline.svg';
+import Icon from './icon.inline.svg';
 import edit from './edit';
 import save from './save';
 import block from './block.json';
@@ -20,8 +20,8 @@ const { registerBlockType } = wp.blocks;
  * Register block
  */
 registerBlockType(block.name, {
-	title: __('template-item-title', 'aquamin'),
-	description: __('template-item-desc', 'aquamin'),
+	title: __('template-title', 'aquamin'),
+	description: __('template-desc', 'aquamin'),
 	icon: Icon,
 	edit,
 	save,