+++ includes/cli/templates/block-dynamic/index.php
@@ -5,14 +5,13 @@
 
 // register the block
 register_block_type_from_metadata(
-	AQUAMIN_BLOCKS . '/block-library/_template-block', [
+	dirname( __FILE__ ), [
 		'render_callback' => function( $attributes, $content, $block ) {
 			ob_start();
 			get_template_part(
-				'blocks/block-library/template-slug/markup',
+				'blocks/block-library/' . basename( dirname( __FILE__ ) ) . '/markup',
 				null,
 				[
-					'class_name' => 'foo',
 					'attributes' => $attributes,
 					'content'    => $content,
 					'block'      => $block,
@@ -21,9 +20,4 @@ register_block_type_from_metadata(
 			return ob_get_clean();
 		},
 	]
-);
-
-// register child blocks
-register_block_type_from_metadata(
-	AQUAMIN_BLOCKS . '/block-library/_template-block/template-item-slug'
 );