+++ includes/cli/templates/block/index.js
@@ -8,15 +8,10 @@
 /**
  * Import dependencies
  */
-import Icon from '../../icons/general.inline.svg';
+import Icon from './icon.inline.svg';
 import edit from './edit';
 import save from './save';
 import block from './block.json';
-
-/**
- * Register inner blocks
- */
-import './template-item-slug';
 
 const { __ } = wp.i18n;
 const { registerBlockType } = wp.blocks;