+++ includes/cli/templates/block/index.php
@@ -1,0 +1,7 @@
+<?php
+/**
+ * Gutenberg block setup
+ */
+
+// register the block
+register_block_type_from_metadata( dirname( __FILE__ ) );