+++ includes/cli/templates/common/script.js
@@ -5,4 +5,5 @@
  * the template-title block.
  */
 
+// replace this demo code with your own:
 console.log('template-slug script has loaded');